+++ arch/vax/include/ansi.h
@@ -1,3 +1,4 @@
+/*	$OpenBSD: ansi.h,v 1.6 1997/07/07 05:56:43 millert Exp $	*/
 /*	$NetBSD: ansi.h,v 1.5 1996/11/15 22:39:01 jtc Exp $	*/
 
 /*-
@@ -49,16 +50,16 @@
  *
  * Thanks, ANSI!
  */
-#define _BSD_CLOCK_T_   unsigned long           /* clock() */
-#define _BSD_PTRDIFF_T_ int                     /* ptr1 - ptr2 */
-#define _BSD_SIZE_T_    unsigned int            /* sizeof() */
-#define _BSD_SSIZE_T_   int                     /* byte count or error */
-#define _BSD_TIME_T_    long                    /* time() */
-#define _BSD_VA_LIST_   char *                  /* va_list */
-#define	_BSD_WCHAR_T_	int                     /* wchar_t */
-#define _BSD_WINT_T_	int			/* wint_t */
-#define _BSD_CLOCKID_T_	int
-#define _BSD_TIMER_T_	int
+#define	_BSD_CLOCK_T_   unsigned long		/* clock() */
+#define	_BSD_PTRDIFF_T_ int			/* ptr1 - ptr2 */
+#define	_BSD_SIZE_T_    unsigned int		/* sizeof() */
+#define	_BSD_SSIZE_T_   int			/* byte count or error */
+#define	_BSD_TIME_T_    int			/* time() */
+#define	_BSD_VA_LIST_   char *			/* va_list */
+#define	_BSD_WCHAR_T_	int			/* wchar_t */
+#define	_BSD_WINT_T_	int			/* wint_t */
+#define	_BSD_CLOCKID_T_	int
+#define	_BSD_TIMER_T_	int
 
 
 #endif  /* _ANSI_H_ */
